+++ server/src/entities/Chat.ts
@@ -1,0 +1,9 @@
+import { gql } from "apollo-server-core";
+
+export const typeDef = gql`
+  type Chat{
+        id: ID!
+        team: Team!
+        messages: [Message]!
+    }
+`
