+++ server/src/entities/Team.ts
@@ -1,0 +1,13 @@
+import { gql } from "apollo-server-core";
+
+export const typeDef = gql`
+    type Team{
+        id: ID!
+        name:String!
+        teamId:String!
+        game: Game!
+        coverUrl: String!
+        users: [Profile]!
+        chat: Chat
+    }
+`
