+++ server/src/schemas/schema.ts
@@ -66,137 +66,12 @@ export const schema = gql `
 
     # }
     
-    type Game{
-        id:ID!
-        name:String!
-        gameId:String!
-        maxSize: Int!
-        coverUrl: String!
-        teams: [Team]!
-        users: [Profile]!
-    }
-
-    type Team{
-        id: ID!
-        name:String!
-        teamId:String!
-        game: Game!
-        coverUrl: String!
-        users: [Profile]!
-        chat: Chat
-    }
-
-    type Chat{
-        id: ID!
-        team: Team!
-        messages: [Message]!
-    }
-
-    
-    type Message{
-        id:ID!
-        text: String!
-        sender: Profile!
-        chat: Chat!
-        sentAt: DateTime!
-    }
-
-
-
-    type AuthPayload{
-        user: User!
-        auth_token: String!
-    }
-
-    type Device{
-        id: ID!
-        deviceId: ID!
-        createdAt: DateTime!
-    }
-    
-    # accessible using id
-    type User{
-        id: ID!
-        userId: ID @isAuth
-        idToken: String @isAuth
-        email: String @isAuth
-        password: String @isAuth
-        authStage: AuthStage!
-        authType: AuthType!
-        createdAt: DateTime!
-        dob: DateTime @isAuth
-        device: Device!
-        profile: Profile!
-    }
-
    
-    enum AuthType{
-        GOOGLE
-        FACEBOOK
-        EMAIL
-    }
-
-    enum AuthStage{
-        SIGNUP
-        USERNAME
-        SUB
-        FINAL
-    }
 
 
     # accessible using username
-    type Profile{
-        id: ID!
-        name: String!
-        username: String
-        avatarUrl: String!
-        bio: String
-        user: User @isAuth
-        games: [Game]!
-        teams: [Team]!
-    }
-
-    type SetUsernamePayload{
-        username: String!,
-        profile_id: ID!
-
-    }
-
-    input SetUsername{
-        username: String!,
-        profile_id: ID!
-
-    }
-    input GoogleUserInput{
-        userId: ID!
-        name: String!
-        email: String!
-        dob: DateTime!
-        avatarUrl: String!
-    }
-
-    input UserInputSignUp{
-        userId: ID!
-        name: String!
-        email: String!
-        avatarUrl: String!
-        authType: AuthType
-        dID: ID!
-    }
-
-    input GameInput{
-        name: String!
-        gameId:String!
-        maxSize: Int!
-        coverUrl: String!
-    }
-
-    input TeamInput{
-        name: String!
-        teamId:String!
-        coverUrl: String!
-        gId: String!
-    }
+   
+    
 
 `
